+++ src/formFields/greenhouse/index.ts
@@ -1,4 +1,5 @@
 
+import { Sections } from './Sections'
 import { AddressSearchField } from './AddressSearchField'
 import { BasicSelect } from './BasicSelect'
 import { MultiCheckbox } from './MultiCheckBox'
@@ -6,14 +7,24 @@ import { SimpleDropdown } from './SimpleDropdown'
 import { SingleFileUpload } from './SingleFileUpload'
 import { Textarea } from './Textarea'
 import { TextField } from './TextField'
+import { MonthYear } from './MonthYear'
+import { DropdownSearchable } from './DropdownSearchable'
+import { DropdownMulti } from './DropdownMulti'
+
+const inputs = [
+  TextField,
+  SingleFileUpload,
+  AddressSearchField,
+  SimpleDropdown,
+  Textarea,
+  BasicSelect,
+  MultiCheckbox,
+  MonthYear,
+  DropdownSearchable,
+  DropdownMulti,
+]
+
 export const RegisterInputs = async (node: Node = document) => {
-  Promise.all([
-    TextField.autoDiscover(node),
-    SingleFileUpload.autoDiscover(node),
-    AddressSearchField.autoDiscover(node),
-    SimpleDropdown.autoDiscover(node),
-    Textarea.autoDiscover(node),
-    BasicSelect.autoDiscover(node),
-    MultiCheckbox.autoDiscover(node),
-  ])
+  await Sections.autoDiscover(node)
+  Promise.all(inputs.map((i) => i.autoDiscover(node)))
 }