+++ src/formFields/greenhouse/xpaths.ts
@@ -16,6 +16,21 @@ export const xpaths = {
   SIMPLE_DROPDOWN: [
     ".//div[@class='field']",
     "[.//div[contains(@class, 'select2-container')]]",
+    "[not(.//div[contains(@class, 'select2-container-multi')])]",
+    `[.//select]`,
+  ].join(''),
+
+
+  DROPDOWN_MULTI: [
+    ".//div[@class='field']",
+    "[.//div[contains(@class, 'select2-container-multi')]]",
+    `[.//select]`,
+  ].join(''),
+
+  DROPDOWN_SEARCHABLE: [
+    ".//div[@class='field']",
+    "[.//div[contains(@class, 'select2-container')]]",
+    `[not(.//select)]`,
   ].join(''),
 
   BASIC_SELECT: [
@@ -29,5 +44,11 @@ export const xpaths = {
   MULTI_CHECKBOX: [
     ".//div[starts-with(@class, 'field')]",
     "[count(.//input[@type='checkbox']) > 1]",
+  ].join(''),
+
+  MONTH_YEAR: [
+    ".//div[@class='field']",
+    `[.//input[@type="text"][@placeholder="MM"]]`,
+    `[.//input[@type="text"][@placeholder="YYYY"]]`,
   ].join(''),
 }